todo-a修正完了ver3  9/20
