+++ src/index.css
@@ -134,7 +134,7 @@
   font-size: 10px !important;
   height: 20px !important;
   min-width: 20px !important;
-  width: auto !important;  /* ← これを追加 */
+  width: auto !important;
   border: none !important;
   border-radius: 3px !important;
   cursor: pointer !important;
