+++ src/todos/components/todos/index.tsx
@@ -1,19 +1,7 @@
 import React, { useState, useEffect, useCallback, useMemo } from 'react';
-
-type Todo = {
-  title: string;
-  readonly id: number;
-  completed_flg: boolean;
-  delete_flg: boolean;
-  progress: number;
-  start_date: string;
-  due_date: string;
-  description: string;
-  start_note: string;
-  images?: { name: string; url: string }[];
-};
-
-type Filter = 'all' | 'completed' | 'unchecked' | 'delete';
+import { Todo, Filter } from '../../types';
+import TodoItemComponent from '../TodoItem';
+import MarkdownRenderer from '../MarkdownRenderer';
 
 const TodoApp: React.FC = () => {
   const [currentView, setCurrentView] = useState<string>('calendar');
@@ -65,538 +53,6 @@ const TodoApp: React.FC = () => {
     setImageViewMode(false);
     setShowMarkdown(true);
   }, []);
-
-  interface MarkdownRendererProps {
-    description: string;
-    onBack: () => void;
-    images?: { name: string; url: string }[];
-  }
-
-  const MarkdownRenderer: React.FC<MarkdownRendererProps> = ({ description, onBack, images = [] }) => {
-    const handleImageClick = (imageUrl: string) => {
-      handleImageView(imageUrl);
-    };
-
-    const renderDescription = (description: string): React.ReactNode => {
-      if (!description || typeof description !== 'string' || description.trim() === '') {
-        return null;
-      }
-      
-      const lines = description.split('\n');
-      const elements: React.ReactNode[] = [];
-      let codeBlockContent: string[] = [];
-      let inCodeBlock = false;
-      let listItems: { bullet: string; text: string; indent: string }[] = [];
-      
-      const flushList = () => {
-        if (listItems.length > 0) {
-          elements.push(
-            <div key={`list-${elements.length}`} style={{ marginBottom: '15px' }}>
-              {listItems.map((item, idx) => (
-                <div key={idx} style={{ 
-                  display: 'flex',
-                  alignItems: 'flex-start',
-                  marginBottom: '5px',
-                  marginLeft: item.indent,
-                  fontSize: '14px',
-                  lineHeight: '1.5'
-                }}>
-                  <span style={{ 
-                    marginRight: '8px',
-                    color: '#333',
-                    minWidth: '16px',
-                    fontSize: '14px',
-                    fontFamily: 'monospace'
-                  }}>
-                    {item.bullet}
-                  </span>
-                  <span 
-                    style={{ color: '#333' }}
-                    dangerouslySetInnerHTML={{ __html: item.text }}
-                  />
-                </div>
-              ))}
-            </div>
-          );
-          listItems = [];
-        }
-      };
-      
-      lines.forEach((line, index) => {
-        // コードブロック処理
-        if (line.trim().startsWith('```')) {
-          flushList();
-          if (inCodeBlock) {
-            elements.push(
-              <pre key={`code-${index}`} style={{ 
-                backgroundColor: '#f5f5f5',
-                border: '1px solid #ddd',
-                borderRadius: '4px',
-                padding: '12px',
-                fontSize: '13px',
-                fontFamily: '"Courier New", Consolas, monospace',
-                overflow: 'auto',
-                margin: '15px 0',
-                lineHeight: '1.4'
-              }}>
-                <code>{codeBlockContent.join('\n')}</code>
-              </pre>
-            );
-            codeBlockContent = [];
-            inCodeBlock = false;
-          } else {
-            inCodeBlock = true;
-          }
-          return;
-        }
-        
-        if (inCodeBlock) {
-          codeBlockContent.push(line);
-          return;
-        }
-        
-        // セキュリティガイドライン行の処理（最優先）
-        if (line.trim().includes('セキュリティガイドライン')) {
-          flushList();
-          
-          // 注意ヘッダー表示
-          elements.push(
-            <div 
-              key={`security-title-${index}`} 
-              style={{ 
-                fontSize: '16px',
-                fontWeight: 'bold',
-                color: '#333',
-                marginBottom: '12px',
-                marginTop: '20px',
-                lineHeight: '1.6',
-                display: 'block',
-                width: '100%',
-                backgroundColor: '#f0f0f0',
-                padding: '8px 12px',
-                borderRadius: '4px',
-                border: '1px solid #ddd'
-              }}
-            >
-              注意
-            </div>
-          );
-          
-          // セキュリティガイドライン内容表示
-          elements.push(
-            <div 
-              key={`security-content-${index}`} 
-              style={{ 
-                fontSize: '14px',
-                color: '#333',
-                marginBottom: '8px',
-                lineHeight: '1.6',
-                display: 'block',
-                width: '100%'
-              }}
-            >
-              :
-              <a 
-                href="https://example.com" 
-                target="_blank" 
-                rel="noopener noreferrer" 
-                style={{
-                  color: '#0066cc', 
-                  textDecoration: 'underline',
-                  marginLeft: '0px',
-                  marginRight: '0px'
-                }}
-              >
-                セキュリティーガイドライン
-              </a>
-              に準拠すること
-            </div>
-          );
-          return;
-        }
-        
-        // 空行処理
-        if (line.trim() === '') {
-          flushList();
-          if (elements.length > 0) {
-            elements.push(
-              <div key={`space-${index}`} style={{ marginBottom: '12px' }} />
-            );
-          }
-        }
-        // 注意行の処理（最優先で処理）- インラインコード形式にも対応
-        else if (line.trim().startsWith('注意：') || line.trim().startsWith('`注意`：') || line.trim().startsWith('`注意`:')) {
-          flushList();
-          let content = '';
-          
-          // 様々なパターンに対応
-          if (line.trim().startsWith('注意：')) {
-            content = line.replace(/^\s*注意：\s*/, '');
-          } else if (line.trim().startsWith('`注意`：')) {
-            content = line.replace(/^\s*`注意`：\s*/, '');
-          } else if (line.trim().startsWith('`注意`:')) {
-            content = line.replace(/^\s*`注意`:\s*/, '');
-          }
-          
-          // 注意を###見出しと同じスタイルで表示（灰色背景付き）
-          elements.push(
-            <div 
-              key={`notice-title-${index}`} 
-              style={{ 
-                fontSize: '16px',
-                fontWeight: 'bold',
-                color: '#333',
-                marginBottom: '12px',
-                marginTop: '20px',
-                lineHeight: '1.6',
-                display: 'block',
-                width: '100%',
-                backgroundColor: '#f0f0f0',
-                padding: '8px 12px',
-                borderRadius: '4px',
-                border: '1px solid #ddd'
-              }}
-            >
-              注意
-            </div>
-          );
-          
-          // 内容を別の行として表示
-          if (content) {
-            let processedContent = content;
-            
-            // 太字処理
-            processedContent = processedContent.replace(/\*\*([^*]+)\*\*/g, '<strong>$1</strong>');
-            
-            // 斜体処理（太字でない場合のみ）
-            processedContent = processedContent.replace(/(?<!\*)\*([^*]+)\*(?!\*)/g, '<em>$1</em>');
-            
-            // 取り消し線
-            processedContent = processedContent.replace(/~~([^~]+)~~/g, '<del style="color: #888;">$1</del>');
-            
-            // インラインコード
-            processedContent = processedContent.replace(/`([^`]+)`/g, 
-              '<code style="background-color: #f5f5f5; padding: 2px 6px; border-radius: 3px; font-family: Consolas, Monaco, \'Courier New\', monospace; font-size: 13px; border: 1px solid #e1e1e1;">$1</code>'
-            );
-            
-            // リンク処理（リンクテキストとURLを適切に処理）
-            processedContent = processedContent.replace(/\[([^\]]+)\]\(([^)]+)\)/g, '<a href="$2" target="_blank" rel="noopener noreferrer" style="color: #0066cc; text-decoration: underline;">$1</a>');
-            
-            elements.push(
-              <div 
-                key={`notice-content-${index}`} 
-                style={{ 
-                  fontSize: '14px',
-                  color: '#333',
-                  marginBottom: '8px',
-                  lineHeight: '1.6',
-                  display: 'block',
-                  width: '100%'
-                }}
-                dangerouslySetInnerHTML={{ __html: processedContent }}
-              />
-            );
-          }
-        }
-        // 見出し処理（###を最優先でチェック）
-        else if (line.startsWith('### ')) {
-          flushList();
-          const headingText = line.substring(4);
-          
-          elements.push(
-            <div 
-              key={`h3-${index}`} 
-              style={{ 
-                fontSize: '16px',
-                fontWeight: 'bold',
-                color: '#333',
-                marginBottom: '12px',
-                marginTop: '20px',
-                lineHeight: '1.6'
-              }}
-            >
-              {headingText}
-            </div>
-          );
-          
-          // 「実装項目」または「進捗」の場合は水平線を追加
-          if (headingText === '実装項目' || headingText === '進捗') {
-            elements.push(
-              <hr 
-                key={`hr-after-${index}`} 
-                style={{ 
-                  margin: '10px 0 15px 0',
-                  border: 'none',
-                  borderTop: '1px solid #ddd'
-                }} 
-              />
-            );
-          }
-        }
-        // ##の処理
-        else if (line.startsWith('## ')) {
-          flushList();
-          elements.push(
-            <div 
-              key={`h2-${index}`} 
-              style={{ 
-                fontSize: '18px',
-                fontWeight: 'bold',
-                color: '#333',
-                marginBottom: '15px',
-                marginTop: '25px',
-                lineHeight: '1.6'
-              }}
-            >
-              {line.substring(3)}
-            </div>
-          );
-        }
-        // #の処理
-        else if (line.startsWith('# ')) {
-          flushList();
-          elements.push(
-            <div 
-              key={`h1-${index}`} 
-              style={{ 
-                fontSize: '20px',
-                fontWeight: 'bold',
-                color: '#333',
-                marginBottom: '18px',
-                marginTop: '30px',
-                lineHeight: '1.6'
-              }}
-            >
-              {line.substring(2)}
-            </div>
-          );
-        }
-        // チェックボックスリスト処理
-        else if (line.match(/^\s*-\s*\[[x ]\]/)) {
-          flushList();
-          const isChecked = line.includes('[x]');
-          const indent = (line.match(/^\s*/)?.[0].length || 0) * 10;
-          const text = line.replace(/^\s*-\s*\[[x ]\]\s*/, '');
-          
-          elements.push(
-            <div key={`checkbox-${index}`} style={{ 
-              display: 'flex',
-              alignItems: 'center',
-              marginBottom: '8px',
-              marginLeft: `${indent}px`,
-              fontSize: '14px',
-              lineHeight: '1.5'
-            }}>
-              <span style={{ 
-                marginRight: '8px',
-                color: '#333',
-                minWidth: '16px',
-                fontSize: '14px',
-                fontFamily: 'monospace'
-              }}>
-                ・
-              </span>
-              <input 
-                type="checkbox" 
-                checked={isChecked} 
-                readOnly 
-                style={{ 
-                  marginRight: '8px',
-                  cursor: 'default'
-                }}
-              />
-              <span style={{ 
-                color: isChecked ? '#888' : '#333'
-              }}>
-                {text}
-              </span>
-            </div>
-          );
-        }
-        // 通常のリスト処理（- または *で始まる）
-        else if (line.match(/^\s*[-*]\s+/)) {
-          const indent = (line.match(/^\s*/)?.[0].length || 0) * 12;
-          let text = line.replace(/^\s*[-*]\s+/, '');
-          
-          // リスト項目内でも取り消し線処理を適用
-          text = text.replace(/~~([^~]+)~~/g, '<del style="color: #888;">$1</del>');
-          
-          listItems.push({
-            bullet: '・',
-            text: text,
-            indent: `${indent}px`
-          });
-        }
-        // 引用処理
-        else if (line.startsWith('> ')) {
-          flushList();
-          elements.push(
-            <div 
-              key={`quote-${index}`} 
-              style={{ 
-                fontSize: '14px', 
-                marginBottom: '8px',
-                lineHeight: '1.6',
-                color: '#333',
-                marginLeft: '2em'
-              }}
-            >
-              {line.substring(2)}
-            </div>
-          );
-        }
-        // 水平線処理
-        else if (line.trim() === '---') {
-          flushList();
-          elements.push(
-            <hr key={`hr-${index}`} style={{ 
-              margin: '20px 0',
-              border: 'none',
-              borderTop: '1px solid #ddd'
-            }} />
-          );
-        }
-        // 通常のテキスト行
-        else if (line.trim() !== '') {
-          flushList();
-          
-          let processedLine = line;
-          
-          // 太字処理
-          processedLine = processedLine.replace(/\*\*([^*]+)\*\*/g, '<strong>$1</strong>');
-          
-          // 斜体処理（太字でない場合のみ）
-          processedLine = processedLine.replace(/(?<!\*)\*([^*]+)\*(?!\*)/g, '<em>$1</em>');
-          
-          // 取り消し線
-          processedLine = processedLine.replace(/~~([^~]+)~~/g, '<del style="color: #888;">$1</del>');
-          
-          // インラインコード
-          processedLine = processedLine.replace(/`([^`]+)`/g, 
-            '<code style="background-color: #f5f5f5; padding: 2px 6px; border-radius: 3px; font-family: Consolas, Monaco, \'Courier New\', monospace; font-size: 13px; border: 1px solid #e1e1e1;">$1</code>'
-          );
-          
-          // リンク処理（リンクテキストとURLを適切に処理）
-          processedLine = processedLine.replace(/\[([^\]]+)\]\(([^)]+)\)/g, '<a href="$2" target="_blank" rel="noopener noreferrer" style="color: #0066cc; text-decoration: underline;">$1</a>');
-          
-          elements.push(
-            <div 
-              key={`text-${index}`} 
-              style={{ 
-                fontSize: '14px', 
-                marginBottom: '8px',
-                lineHeight: '1.6',
-                color: '#333'
-              }}
-              dangerouslySetInnerHTML={{ __html: processedLine }}
-            />
-          );
-        }
-      });
-      
-      // 残りのリストを処理
-      flushList();
-      
-      return elements;
-    };
-
-    return (
-      <div style={{ 
-        backgroundColor: 'white', 
-        padding: '30px', 
-        borderRadius: '8px', 
-        border: '1px solid #ddd',
-        boxShadow: '0 2px 4px rgba(0,0,0,0.1)'
-      }}>
-        {/* タスクタイトル表示 */}
-        <div style={{
-          fontSize: '20px',
-          fontWeight: 'bold',
-          color: '#333',
-          marginBottom: '20px',
-          padding: '10px 15px',
-          backgroundColor: '#f8f9fa',
-          borderRadius: '6px',
-          border: '1px solid #e9ecef'
-        }}>
-          {markdownTodo?.title || 'タスク詳細'}
-        </div>
-        
-        {renderDescription(description)}
-        
-        {/* 画像表示セクション */}
-        {images && images.length > 0 && (
-          <div style={{ marginTop: '30px' }}>
-            <div style={{ 
-              fontSize: '16px',
-              fontWeight: 'bold',
-              color: '#333',
-              marginBottom: '15px',
-              lineHeight: '1.6'
-            }}>
-              添付画像
-            </div>
-            <div style={{ display: 'flex', flexDirection: 'column', gap: '30px' }}>
-              {images.map((image, index) => (
-                <div key={index} style={{ textAlign: 'center' }}>
-                  <img
-                    src={image.url}
-                    alt={image.name}
-                    onClick={() => handleImageClick(image.url)}
-                    style={{
-                      width: '100%',
-                      maxWidth: '800px',
-                      height: 'auto',
-                      minHeight: '400px',
-                      objectFit: 'contain',
-                      borderRadius: '8px',
-                      border: '1px solid #ddd',
-                      boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
-                      display: 'block',
-                      margin: '0 auto',
-                      cursor: 'pointer',
-                      transition: 'transform 0.2s ease'
-                    }}
-                    onMouseEnter={(e) => {
-                      e.currentTarget.style.transform = 'scale(1.02)';
-                    }}
-                    onMouseLeave={(e) => {
-                      e.currentTarget.style.transform = 'scale(1)';
-                    }}
-                  />
-                  <div style={{
-                    marginTop: '12px',
-                    fontSize: '14px',
-                    color: '#666',
-                    wordBreak: 'break-all'
-                  }}>
-                    {image.name} - クリックで拡大
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-        
-        <div style={{ marginTop: '30px', textAlign: 'left' }}>
-          <button
-            onClick={onBack}
-            style={{
-              padding: '8px 16px',
-              backgroundColor: '#dc3545',
-              color: 'white',
-              border: 'none',
-              borderRadius: '4px',
-              cursor: 'pointer',
-              fontSize: '14px'
-            }}
-          >
-            閉じる
-          </button>
-        </div>
-      </div>
-    );
-  };
 
   const ImageViewPage: React.FC = () => {
     return (
@@ -765,275 +221,6 @@ const TodoApp: React.FC = () => {
           )}
         </div>
       </div>
-    );
-  };
-
-  interface TodoItemComponentProps {
-    todo: Todo;
-    onUpdateTodo: (id: number, key: keyof Todo, value: any) => void;
-    onToggleExpanded: (id: number) => void;
-    isExpanded: boolean;
-    onMarkClick: (todo: Todo) => void;
-  }
-
-  const TodoItemComponent: React.FC<TodoItemComponentProps> = ({ todo, onUpdateTodo, onToggleExpanded, isExpanded, onMarkClick }) => {
-    const [localDescription, setLocalDescription] = useState(todo.description || '');
-    const [dragOver, setDragOver] = useState(false);
-
-    useEffect(() => {
-      setLocalDescription(todo.description || '');
-    }, [todo.description]);
-
-    const handleDescriptionBlur = () => {
-      onUpdateTodo(todo.id, 'description', localDescription);
-    };
-
-    const handleMarkClick = () => {
-      onMarkClick(todo);
-    };
-
-    // ドラッグ&ドロップ処理
-    const handleDragOver = (e: React.DragEvent) => {
-      e.preventDefault();
-      setDragOver(true);
-    };
-
-    const handleDragLeave = (e: React.DragEvent) => {
-      e.preventDefault();
-      setDragOver(false);
-    };
-
-    const handleDrop = (e: React.DragEvent) => {
-      e.preventDefault();
-      setDragOver(false);
-      
-      const files = e.dataTransfer.files;
-      if (files.length > 0) {
-        const file = files[0];
-        if (file.type.startsWith('image/')) {
-          const reader = new FileReader();
-          reader.onload = (event) => {
-            if (event.target?.result) {
-              const imageUrl = event.target.result as string;
-              const imageName = file.name;
-              
-              // 画像情報をTodoに追加
-              const currentImages = todo.images || [];
-              const newImages = [...currentImages, { name: imageName, url: imageUrl }];
-              onUpdateTodo(todo.id, 'images', newImages);
-            }
-          };
-          reader.readAsDataURL(file);
-        }
-      }
-    };
-
-    return (
-      <li style={{ marginBottom: '15px', padding: '15px', border: '1px solid #ddd', borderRadius: '8px', backgroundColor: '#f5deb3' }}>
-        <div style={{ display: 'flex', gap: '20px', minHeight: '80px', alignItems: 'flex-start', fontSize: '14px', color: '#666' }}>
-          <div style={{ flex: '0 0 150px', display: 'flex', gap: '10px', backgroundColor: '#f5deb3', padding: '10px', borderRadius: '5px' }}>
-            <div style={{ flex: '0 0 66px', paddingTop: '15px' }}>
-              <label style={{ display: 'block', marginBottom: '5px', fontWeight: 'bold' }}>進捗率</label>
-              <select
-                value={todo.progress}
-                onChange={(e) => onUpdateTodo(todo.id, 'progress', Number(e.target.value))}
-                disabled={todo.delete_flg}
-                style={{ 
-                  width: '100%', 
-                  padding: '6px', 
-                  fontSize: '12px', 
-                  border: '1px solid #ccc', 
-                  borderRadius: '5px',
-                  backgroundColor: todo.delete_flg ? '#f5f5f5' : 'white'
-                }}
-              >
-                <option value={0}>0%</option>
-                <option value={10}>10%</option>
-                <option value={20}>20%</option>
-                <option value={30}>30%</option>
-                <option value={40}>40%</option>
-                <option value={50}>50%</option>
-                <option value={60}>60%</option>
-                <option value={70}>70%</option>
-                <option value={80}>80%</option>
-                <option value={90}>90%</option>
-                <option value={100}>100%</option>
-              </select>
-            </div>
-            
-            <div style={{ flex: '0 0 105px', display: 'flex', flexDirection: 'column', gap: '10px' }}>
-              <div>
-                <strong>開始日:</strong>
-                <input
-                  type="date"
-                  value={todo.start_date}
-                  onChange={(e) => onUpdateTodo(todo.id, 'start_date', e.target.value)}
-                  disabled={todo.delete_flg}
-                  max={todo.due_date || undefined}
-                  style={{ 
-                    marginLeft: '5px', 
-                    padding: '4px 8px', 
-                    fontSize: '11px', 
-                    border: '1px solid #ccc', 
-                    borderRadius: '4px',
-                    backgroundColor: todo.delete_flg ? '#f5f5f5' : 'white',
-                    width: '95px',
-                    fontFamily: 'Arial, sans-serif',
-                    cursor: todo.delete_flg ? 'not-allowed' : 'pointer'
-                  }}
-                />
-              </div>
-              
-              <div>
-                <strong>完了予定:</strong>
-                <input
-                  type="date"
-                  value={todo.due_date}
-                  onChange={(e) => onUpdateTodo(todo.id, 'due_date', e.target.value)}
-                  disabled={todo.delete_flg}
-                  min={todo.start_date || undefined}
-                  style={{ 
-                    marginLeft: '5px', 
-                    padding: '4px 8px', 
-                    fontSize: '11px', 
-                    border: '1px solid #ccc', 
-                    borderRadius: '4px',
-                    backgroundColor: todo.delete_flg ? '#f5f5f5' : 'white',
-                    width: '95px',
-                    fontFamily: 'Arial, sans-serif',
-                    cursor: todo.delete_flg ? 'not-allowed' : 'pointer'
-                  }}
-                />
-              </div>
-            </div>
-          </div>
-          
-          <div style={{ flex: '1', paddingTop: '55px', display: 'flex', gap: '8px', alignItems: 'flex-end' }}>
-            <div style={{ flex: '1' }}>
-              <input
-                type="text"
-                value={todo.title}
-                onChange={(e) => onUpdateTodo(todo.id, 'title', e.target.value)}
-                disabled={todo.delete_flg || todo.progress === 100}
-                style={{ 
-                  width: 'calc(100% - 10px)', 
-                  padding: '6px', 
-                  fontSize: '12px', 
-                  border: '1px solid #ccc', 
-                  borderRadius: '5px',
-                  backgroundColor: (todo.delete_flg || todo.progress === 100) ? '#f5f5f5' : 'white'
-                }}
-              />
-            </div>
-            <button
-              type="button"
-              onClick={() => onToggleExpanded(todo.id)}
-              className="edit-button"
-              style={{
-                padding: '6px 12px',
-                fontSize: '12px',
-                minWidth: '50px'
-              }}
-            >
-              編集
-            </button>
-            <button 
-              onClick={() => onUpdateTodo(todo.id, 'delete_flg', !todo.delete_flg)}
-              className={todo.delete_flg ? "restore-button" : "delete-button"}
-              style={{
-                padding: '6px 12px',
-                fontSize: '12px',
-                minWidth: '50px'
-              }}
-            >
-              {todo.delete_flg ? '復元' : '削除'}
-            </button>
-          </div>
-        </div>
-        
-        {isExpanded && (
-          <div style={{ marginTop: '15px' }}>
-            <label style={{ display: 'block', marginBottom: '5px', fontWeight: 'bold', fontSize: '14px' }}>詳細説明</label>
-            <div
-              onDragOver={handleDragOver}
-              onDragLeave={handleDragLeave}
-              onDrop={handleDrop}
-              style={{
-                border: dragOver ? '2px solid #007bff' : '1px solid #ccc',
-                borderRadius: '5px',
-                backgroundColor: dragOver ? '#f0f8ff' : ((todo.delete_flg || todo.progress === 100) ? '#f5f5f5' : 'white'),
-                position: 'relative'
-              }}
-            >
-              <textarea
-                value={localDescription}
-                onChange={(e) => setLocalDescription(e.target.value)}
-                onBlur={handleDescriptionBlur}
-                disabled={todo.delete_flg || todo.progress === 100}
-                rows={3}
-                style={{ 
-                  width: '100%', 
-                  padding: '8px', 
-                  fontSize: '14px', 
-                  border: 'none',
-                  borderRadius: '5px',
-                  resize: 'vertical',
-                  fontFamily: 'Arial, sans-serif',
-                  backgroundColor: 'transparent',
-                  outline: 'none'
-                }}
-                placeholder="詳細説明を入力... (画像もここにドラッグ&ドロップ可能)"
-              />
-              {dragOver && (
-                <div style={{
-                  position: 'absolute',
-                  top: '0',
-                  left: '0',
-                  right: '0',
-                  bottom: '0',
-                  backgroundColor: 'rgba(0, 123, 255, 0.1)',
-                  border: '2px dashed #007bff',
-                  borderRadius: '5px',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  pointerEvents: 'none',
-                  zIndex: 1
-                }}>
-                  <div style={{
-                    backgroundColor: 'rgba(0, 123, 255, 0.9)',
-                    color: 'white',
-                    padding: '10px 15px',
-                    borderRadius: '5px',
-                    fontSize: '14px'
-                  }}>
-                    画像をドロップしてください
-                  </div>
-                </div>
-              )}
-            </div>
-            
-            <div style={{ marginTop: '10px', textAlign: 'left' }}>
-              <button
-                type="button"
-                onClick={handleMarkClick}
-                style={{
-                  padding: '6px 12px',
-                  fontSize: '12px',
-                  border: '1px solid #28a745',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  backgroundColor: todo.delete_flg ? '#f5f5f5' : '#28a745',
-                  color: todo.delete_flg ? '#999' : 'white'
-                }}
-                disabled={todo.delete_flg}
-              >
-                マークダウン表示
-              </button>
-            </div>
-          </div>
-        )}
-      </li>
     );
   };
 
@@ -1625,7 +812,9 @@ const TodoApp: React.FC = () => {
                 <MarkdownRenderer 
                   description={markdownTodo.description || ''}
                   images={markdownTodo.images || []}
+                  markdownTodo={markdownTodo}
                   onBack={handleCloseMarkdown}
+                  onImageView={handleImageView}
                 />
               </div>
             </div>
